task4/Study_Notes_summerizer/requirements.txt
streamlit
pypdf
python-dotenv
google-generativeai